backend/requirements.txt
aiofiles==23.2.1
altair==5.3.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.9.1
attrs==23.2.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.12.3
billiard==4.2.1
black==23.7.0
boto3==1.40.35
botocore==1.40.35
cachetools==5.5.2
celery==5.5.3
certifi==2024.2.2
cffi==2.0.0
channels==4.2.2
channels_redis==4.3.0
charset-normalizer==3.3.2
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
contourpy==1.2.1
cryptography==46.0.1
cycler==0.12.1
daphne==4.2.1
distro==1.9.0
dj-database-url==2.2.0
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gradio==4.37.2
gradio_client==1.0.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.23.4
hyperlink==21.0.0
idna==3.7
importlib_resources==6.4.0
incremental==24.7.2
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.5.4
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
msgpack==1.1.1
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.1
openai==1.93.0
openpyxl==3.1.2
orjson==3.10.6
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.2
pathspec==0.11.2
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==3.10.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.32.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.18.0
pyOpenSSL==25.3.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.1
redis==6.2.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9.1
ruff==0.5.1
s3transfer==0.14.0
selenium==4.19.0
semantic-version==2.10.0
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
starlette==0.37.2
tabulate==0.9.0
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
Twisted==25.5.0
txaio==25.6.1
typer==0.12.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
whitenoise==6.7.0
wsproto==1.2.0
zope.interface==8.0
